fix: improve error handling of redirect handler vs error responses
the repl/exec redirect handler was treating error responses as an unhandled redirect type; this is correct at some level, but not very useful to the user. let's do what bash does here, and report the original error!
